A patch from the tracker: Build fix for NetBSD/amd64
Found here: https://sourceforge.net/tracker/index.php?func=detail&aid=2517113&group_id=5776&atid=305776

Here is the original Patch:
http://www.NetBSD.org/cgi-bin/query-pr-single.pl?number=33769

The mjpegtools fail to build because: assembler errors due to a wrong CPU detection. mjpegtools assumes that on an amd64 platform __LP64___ is defined, which turns out not to be the case.  There appears to be some confusion as to whether code should check for __LP64__ or _LP64.

Posted by Thomas Klausner